Moved headers to new directorys Separated nodes into separate headers Made functions for repeated actions in AST generation Added the words "In the name of God, the Gracious, the Merciful" and "بِسۡمِ ٱللَّهِ ٱلرَّحۡمَٰنِ ٱلرَّحِيمِ" to rendorc.cpp and rendorvm.cpp
