+++ Test Files/Hello World.Cren
@@ -1,10 +1,8 @@
 LOAD 0
 DEFINE main
-CONST INT64, 11
+CONST ARITHMETHIC,  1 2 5 + *
 ASSIGN arithmethic
 CONST STRING, Hello
-FINALIZE_CALL echo
-CONST STRING, bar
 FINALIZE_CALL echo
 CONST REFERENCE, arithmethic
 CONST INT64, 10
@@ -20,7 +18,6 @@ CONST REFERENCE, bar
 FINALIZE_CALL echo
 FUNCTION END
 END 0
-
