+++ src/RendorCompiler/InspectAndOptimize/ASTInspectionFunctions.cpp
@@ -1,4 +1,4 @@
-#include "RendorCompiler/Parser.hpp"
+#include "RendorCompiler/Parser/Parser.hpp"
 
 void Parser::InspectTypesReferences(const NodeType& Type, const NodeObject& Node)
 {
