+++ src/RendorCompiler/Logic/MathEvaluator.cpp
@@ -1,4 +1,4 @@
-#include "RendorCompiler/MathEvaluator.hpp"
+#include "RendorCompiler/Math/MathEvaluator.hpp"
 
 static std::string EvalNums (std::string Num1, std::string Num2, std::string Op)
 {
