+++ src/RendorCompiler/Logic/Mathparsing.cpp
@@ -1,4 +1,4 @@
-#include "RendorCompiler/Mathparsing.hpp"
+#include "RendorCompiler/Math/Mathparsing.hpp"
 
 // TODO: finish this lol
 std::string OperationToPostfix (std::string_view Operation){
